+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Version 12.0.0
 - Improve validation for custom field names.
 - Added method ``calver`` for ``Development``.
 - Added method ``stage`` for ``Development``.
+- Added decorator ``@register`` for ``Field`` and ``Fieldset`` to register custom fields.
 - Removed parameter ``providers`` for ``Field`` and ``Fieldset``. Use custom fields instead.
 - Removed parameters ``pre_release`` and ``calver`` for ``Development.version``. Use methods ``stage`` and ``calver`` instead.